treat Timesteps generically with int type for existing schedulers and float for new ones starting with EulerDiscrete

+++ com.doji.diffusers/Runtime/Scripts/Models/Unet.cs
@@ -63,7 +63,7 @@ namespace Doji.AI.Diffusers {
             return _worker.PeekOutput("out_sample") as TensorFloat;
         }
 
-        public Tensor CreateTimestep(TensorShape shape, int t) {
+        public Tensor CreateTimestep(TensorShape shape, float t) {
             if (_model == null) {
                 throw new NullReferenceException($"{nameof(_model)} was null");
             }
